fix: rename file and remove comment
