+++ src/components/AssetManagementPanel/AssetManagementForm/FormVariants/NominateBeneficiaryHolder/index.tsx
@@ -1,1 +1,1 @@
-export * from "./NominateBeneficiaryHolder";
+export * from "./NominateBeneficiary";
